Generate checklist line-list widget from app-interface
