+++ cargo-lens/src/app.rs
@@ -3,13 +3,17 @@ use std::marker::PhantomData;
 use ratatui::{
     backend::Backend,
     layout::{Constraint, Direction, Layout},
-    widgets::{Block, Borders, List, Paragraph},
+    style::{Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, Borders, List, ListItem, Paragraph},
     Frame,
 };
 
 use crate::review_req_checklist::ReviewReqChecklist;
 
+/// Central hub for data/widget reference.
 pub struct App<B> {
+    /// raw-data on checklist-items
     pub list: ReviewReqChecklist,
     _phantom: PhantomData<B>,
 }
@@ -26,7 +30,7 @@ impl<B: Backend> App<B> {
             .direction(Direction::Horizontal)
             .constraints([Constraint::Percentage(30), Constraint::Min(0)].as_ref())
             .split(f.size());
-        let items = self.list.lines();
+        let items = self.lines();
 
         let block = Block::default().title("Checklist").borders(Borders::ALL);
         let checklist = List::new(items).block(block);
@@ -36,5 +40,37 @@ impl<B: Backend> App<B> {
             Paragraph::new::<&str>(self.list.items.get(self.list.index).unwrap().info.as_ref())
                 .block(block);
         f.render_widget(info, chunks[1]);
+    }
+    pub fn lines(&self) -> Vec<ListItem> {
+        let tick = "✓";
+        let cross = "×";
+        let span = |fill| -> Vec<Span> {
+            ["[", fill, "] - "]
+                .into_iter()
+                .map(|st: &str| Span::raw(st))
+                .collect()
+        };
+
+        let lines: Vec<ListItem> = self
+            .list
+            .items
+            .iter()
+            .enumerate()
+            .map(|(i, item)| {
+                let mut spans = if item.toggled {
+                    span(tick)
+                } else {
+                    span(cross)
+                };
+                spans.push(Span::raw(&item.name));
+                let res = ListItem::new(Line::from(spans));
+                if i == self.list.index {
+                    res.style(Style::default().add_modifier(Modifier::BOLD))
+                } else {
+                    res
+                }
+            })
+            .collect();
+        lines
     }
 }
